requirements.txt
streamlit>=1.30.0
streamlit-authenticator>=0.3.1
reportlab>=4.1.0
Pillow>=10.0.0
pyyaml>=6.0
